Backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.31.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.30.0,<0.31.0
psycopg[binary]>=3.1.0,<4.0.0
alembic>=1.13.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=3.2.0,<4.0.0
Pillow>=10.0.0,<11.0.0
PyPDF2>=3.0.0,<4.0.0
reportlab>=4.0.0,<4.2.0
python-multipart>=0.0.9,<1.0.0
httpx>=0.27.0,<0.28.0
